Added zwave component
.gitignore extended with files that shouldn't be committed, notably
`options.xml` that contains the `NetworkKey`.

The `zwave.yml` contains a `device_config_glob` section, which I had
quite some issues with getting to load properly in Home Assistant. It
just didn't want to take the "*2" glob.
I was restarting home assistant to check it was working, but it
continuously was not. Between restarts I removed the
`.storage/core.entity_registry` file to force reloads. But _that_ turned
out to be the problem: for eg the `*_sourcenodeid` glob the entity _does
not_ have to exist in the `core.entity_registry` file, for the `*2` glob
it _does_ need to have appeared in there.
That also means the `_2` entities will always show up as "entity
unavailable" in the zwave configuration panel, but at least it finally
looks the way I want it.